Removed old classes. Moved classes from next package, to default package
