+++ src/main/java/com/github/bgora/rpnlibrary/Calculator.java
@@ -24,8 +24,8 @@ import com.github.bgora.rpnlibrary.exceptions.WrongArgumentException;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
-import java.util.Deque;
-import java.util.LinkedList;
+import java.util.Optional;
+import java.util.function.Function;
 
 /**
  * RPN Calculator Implementation with functions.
@@ -35,11 +35,9 @@ import java.util.LinkedList;
  */
 public class Calculator {
 
-    public static final String ZERO = "0.0";
-    public static final String EMPTY_SPACE = " ";
-    public static final String COMMA = ",";
-    protected final RPNChecking checker;
-    protected final RPNExecuting executioner;
+    protected final Function<String, String> transformer;
+    protected final Function<String, String> rpnFactory;
+    protected final Function<String, BigDecimal> rpnCalculator;
     private final MathContext mathContext;
     private final int SCALE;
 
@@ -55,10 +53,15 @@ public class Calculator {
      * @see RPNExecuting
      */
     public static Calculator createCalculator() {
-        final CalculationEngine calculationEngine = new CalculatorEngine(StrategiesUtil.DEFAULT_OPERATORS,
-                                                                         StrategiesUtil.DEFAULT_FUNCTIONS);
         final MathContext mathContext = MathContext.DECIMAL64;
-        return new Calculator(calculationEngine, calculationEngine, mathContext, 2);
+        final StrategyProvider strategyProvider = new DefaultStrategyProvider();
+        final RPNChecking rpnChecking = new RPNChecker(strategyProvider);
+        final RPNExecuting rpnExecuting = new DefaultRPNExecutor(strategyProvider);
+        return new Calculator(
+                new InputTransformer(rpnChecking),
+                new RPNFactory(rpnChecking),
+                new RPNCalculator(2,rpnChecking, rpnExecuting, mathContext)
+                , mathContext, 2);
     }
 
     /**
@@ -67,9 +70,8 @@ public class Calculator {
      * own operations. To do so, you have to implement you own objects, that
      * implementas {@code RPNChecking}, and
      * {@code RPNExecuting}.
+     * .
      *
-     * @param checker     Object implementing RPNChecking - Used for checking operations in input.
-     * @param executioner Object implementing RPNExecuting - used for executing operations on input.
      * @param mathContext MathContext - Set Rounding Mode, and precision
      * @param scale       scale number of digits after .
      * @return new Instance of {@code pl.bgora.Calculator}
@@ -77,215 +79,39 @@ public class Calculator {
      * @see RPNExecuting
      */
 
-    public static Calculator createCalculator(RPNChecking checker,
-                                              RPNExecuting executioner,
+    public static Calculator createCalculator(Function<String, String> transformer,
+                                              Function<String, String> rpnFactory,
+                                              Function<String, BigDecimal> rpnCalculator,
                                               final MathContext mathContext,
                                               final int scale) {
-        return new Calculator(checker, executioner, mathContext, scale);
+        return new Calculator(transformer, rpnFactory, rpnCalculator, mathContext, scale);
     }
 
     /**
      * Constructor Creates an instance of the class.
      *
-     * @param checker     Object implementing RPNChecking - Used for checking operations in input.
-     * @param executioner Object implementing RPNExecuting - used for executing operations on input.
+     * @param transformer
+     * @param rpnFactory
+     * @param rpnCalculator
      * @param mathContext
      * @param scale
      */
-    private Calculator(RPNChecking checker, RPNExecuting executioner, final MathContext mathContext, final int scale) {
-        this.checker = checker;
-        this.executioner = executioner;
+    private Calculator(final Function<String, String> transformer, final Function<String, String> rpnFactory, final Function<String, BigDecimal> rpnCalculator, final MathContext mathContext, final int scale) {
+        this.transformer = transformer;
+        this.rpnFactory = rpnFactory;
+        this.rpnCalculator = rpnCalculator;
         this.mathContext = mathContext;
         this.SCALE = scale;
     }
 
     public BigDecimal calculate(final String input) throws WrongArgumentException, NoSuchFunctionFound {
-        final String temp = prepareInput(input);
-        final String result = createRPN(temp);
-        return getResult(result);
+        return Optional.of(input)
+                .map(transformer)
+                .map(rpnFactory)
+                .map(rpnCalculator)
+                .orElseThrow(()->new WrongArgumentException("Incorrect input"));
     }
 
-    /**
-     * Format input for further processing.
-     *
-     * @param input Input string.
-     * @return Formatted String.
-     * @throws WrongArgumentException Thrown if the input is incorrect (Incorrect format, or
-     *                                unsupported operations)
-     */
-    private String prepareInput(String input) throws WrongArgumentException {
-        final StringBuilder result = new StringBuilder();
-        final String inputValue = input.trim();
-        int length = inputValue.length();
-        char character;
-        boolean lastWasDigit = false;
-        boolean lastWasOperator = false;
-        boolean lastWasWhiteSpace = false;
-        boolean lastWasLetter = false;
-        for (int i = 0; i < length; i++) {
-            character = inputValue.charAt(i);
-            if (isDigitOrSeparator(character) && (lastWasDigit || !lastWasOperator)) {
-                lastWasDigit = true;
-                result.append(character);
-                lastWasWhiteSpace = false;
-                lastWasLetter = false;
-            } else if (Character.isDigit(character)) {
-                lastWasDigit = true;
-                lastWasLetter = false;
-                lastWasOperator = false;
-                if (!lastWasWhiteSpace) {
-                    result.append(EMPTY_SPACE);
-                }
-                result.append(character);
-                lastWasWhiteSpace = false;
-            } else if (checker.isOperatorOrBracket(String.valueOf(character))) {
-                lastWasDigit = false;
-                lastWasLetter = false;
-                lastWasOperator = true;
-                if (!lastWasWhiteSpace) {
-                    result.append(EMPTY_SPACE);
-                }
-                result.append(character);
-                lastWasWhiteSpace = false;
-            } else if (Character.isWhitespace(character)) {
-                if (!lastWasWhiteSpace && !lastWasDigit) {
-                    result.append(EMPTY_SPACE);
-                    lastWasWhiteSpace = true;
-                }
-                lastWasDigit = false;
-                lastWasOperator = false;
-            } else if (Character.isLetter(character)) {
-                lastWasDigit = false;
-                lastWasOperator = false;
-                if (!lastWasLetter && !lastWasWhiteSpace) {
-                    result.append(EMPTY_SPACE).append(character);
-                } else {
-                    result.append(character);
-                }
-                lastWasWhiteSpace = false;
-                lastWasLetter = true;
-            } else {
-                throw new WrongArgumentException("Element \"" + character + "\" is not recognized by the Checker");
-            }
-        }
-
-        return result.toString().trim();
-    }
-
-    private boolean isDigitOrSeparator(char c) {
-        return Character.isDigit(c) || c == '.' || c == ',';
-    }
-
-    /**
-     * Creates String in Reverse Polish Notation.
-     *
-     * @param input Input String in "Natural" format.
-     * @return String formatted into RPN.
-     * @throws WrongArgumentException Thrown if the input is incorrect (Incorrect format, or
-     *                                unsupported opertians)
-     */
-    private String createRPN(String input) throws WrongArgumentException {
-        final String trimmed = input.trim();
-        final StringBuilder result = new StringBuilder();
-        final Deque<String> stack = new LinkedList<>();
-        final String[] factors = trimmed.split(EMPTY_SPACE);
-        final int length = factors.length;
-        String temp;
-        String stackOperator;
-        for (int i = 0; i < length; i++) {
-            temp = factors[i];
-            if (checker.isDigit(temp)) {
-                result.append(EMPTY_SPACE).append(temp);
-            } else if (checker.isFunction(temp)) {
-                stack.push(temp);
-            } else if (COMMA.equals(temp)) {
-                do {
-                    stackOperator = stack.pop();
-                    if (!checker.isLeftBracket(stackOperator)) {
-                        result.append(EMPTY_SPACE).append(stackOperator);
-                    }
-                } while (!stack.isEmpty() && !checker.isLeftBracket(stackOperator));
-            } else if (checker.isOperator(temp)) {
-                while (!stack.isEmpty() && checker.isOperator(stack.peek())) {
-                    stackOperator = stack.peek();
-                    if (checker.isLeftAssociativity(stackOperator) && (checker.compareOperators(stackOperator,
-                                                                                                temp) >= 0)) {
-                        stack.pop();
-                        result.append(EMPTY_SPACE).append(stackOperator);
-                    } else if (checker.isRightAssociativity(stackOperator)
-                            && (checker.compareOperators(stackOperator, temp) > 0)) {
-                        stack.pop();
-                        result.append(EMPTY_SPACE).append(stackOperator);
-                    } else {
-                        break;
-                    }
-                }
-                stack.push(temp);
-            } else if (checker.isLeftBracket(temp)) {
-                stack.push(temp);
-            } else if (checker.isRightBracket(temp)) {
-                do {
-                    temp = stack.pop();
-                    if (!checker.isLeftBracket(temp)) {
-                        result.append(EMPTY_SPACE).append(temp);
-                    }
-                } while (!checker.isLeftBracket(temp));
-                if (!stack.isEmpty() && checker.isFunction(stack.peek())) {
-                    result.append(EMPTY_SPACE).append(stack.pop());
-                }
-            } else {
-                throw new WrongArgumentException("Element \"" + temp + "\" is not recognized by the Checker");
-            }
-        }
-        while (!stack.isEmpty()) {
-            result.append(EMPTY_SPACE).append(stack.pop());
-        }
-
-        return result.toString().trim();
-    }
-
-    /**
-     * Calculates RPN String into BigDecimal.
-     *
-     * @param result Input RPN String
-     * @return value as {@code java.math.BigDecimal}
-     * @throws WrongArgumentException
-     * @throws NoSuchFunctionFound
-     */
-    private BigDecimal getResult(String result) throws WrongArgumentException, NoSuchFunctionFound {
-        final String[] factors = result.trim().split(EMPTY_SPACE);
-        final Deque<String> stack = new LinkedList<String>();
-        String temp;
-        String variable1;
-        String variable2;
-        BigDecimal value;
-        for (int i = 0; i < factors.length; i++) {
-            temp = factors[i];
-            if (checker.isDigit(temp)) {
-                stack.push(temp);
-            } else if (checker.isOperator(temp)) {
-                variable1 = stack.pop();
-                if (!stack.isEmpty()) {
-                    variable2 = stack.pop();
-                } else {
-                    variable2 = ZERO;
-                }
-                value = executioner.executeOperator(temp, mathContext, variable2, variable1);
-                stack.push(value.toPlainString());
-            } else if (checker.isFunction(temp)) {
-                int count = checker.getFunctionParamsCount(temp);
-                String[] table = new String[count];
-                String params = stack.pop();
-                String[] paramsTable = params.split(COMMA);
-                for (int j = 0; j < count; j++) {
-                    table[j] = paramsTable[j];
-                }
-                value = executioner.executeFunction(temp, mathContext, table);
-                stack.push(value.toPlainString());
-            }
-        }
-        return new BigDecimal(stack.pop()).setScale(SCALE, mathContext.getRoundingMode());
-    }
 
     public MathContext getMathContext() {
         return new MathContext(mathContext.getPrecision(), mathContext.getRoundingMode());
