+++ src/main/java/com/github/bgora/rpnlibrary/InputTransformer.java
@@ -1,4 +1,4 @@
-package com.github.bgora.rpnlibrary.next;
+package com.github.bgora.rpnlibrary;
 
 
 import com.github.bgora.rpnlibrary.exceptions.WrongArgumentException;
