+++ src/main/java/com/github/bgora/rpnlibrary/StrategyProvider.java
@@ -1,4 +1,4 @@
-package com.github.bgora.rpnlibrary.next;
+package com.github.bgora.rpnlibrary;
 
 import com.github.bgora.rpnlibrary.functions.AbstractFunctionStrategy;
 import com.github.bgora.rpnlibrary.operators.AbstractOperatorStrategy;
